Fixed negative id issue, swagger is now at index.

+++ NoodleApi/NoodleApi/Models/Noodle.cs
@@ -9,11 +9,11 @@ namespace NoodleApi.Models
 {
     public class Noodle
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
         public string Name { get; set; }
-        public long BrandId { get; set; }
+        public int BrandId { get; set; }
         public string Flavor { get; set; }
         public string ImgUrl { get; set; }
         public string Description { get; set; }
